+++ src/FPath/RelDir.hs
@@ -2,64 +2,47 @@
 {-# LANGUAGE DerivingStrategies         #-}
 
 module FPath.RelDir
-  ( AsRelDir( _RelDir ), RelDir
+  ( AsRelDir( _RelDir ), RelDir --, AsNonRootRelDir( _NonRootRelDir )
+  , NonRootRelDir
 
-  , reldir, reldirT
+  , reldir, reldirN, reldirT
   , parseRelDirP
 
   , tests
   )
 where
 
+import Base1T  hiding ( toList )
 import Prelude  ( error )
 
 -- base --------------------------------
 
-import Control.Applicative  ( pure )
-import Control.Monad        ( mapM, return )
-import Data.Bool            ( otherwise )
-import Data.Either          ( Either( Left, Right ), either )
-import Data.Eq              ( Eq )
-import Data.Foldable        ( concat, foldl', foldl1, foldMap, foldr, foldr1 )
-import Data.Function        ( (&), ($), const, id )
-import Data.Functor         ( fmap )
-import Data.Maybe           ( Maybe( Just, Nothing ) )
-import Data.Monoid          ( Monoid )
-import Data.Semigroup       ( Semigroup )
-import Data.String          ( String )
-import Data.Typeable        ( Proxy( Proxy ), TypeRep, typeRep )
-import GHC.Exts             ( IsList( fromList, toList ), Item )
-import System.Exit          ( ExitCode )
-import System.IO            ( IO )
-import Text.Show            ( Show( show ) )
-
--- base-unicode-symbols ----------------
-
-import Data.Eq.Unicode        ( (≡) )
-import Data.Function.Unicode  ( (∘) )
-import Data.Monoid.Unicode    ( (⊕) )
+import Data.Foldable  ( concat, foldMap )
+import Data.Monoid    ( Monoid( mempty ) )
+import Data.Typeable  ( Proxy( Proxy ), TypeRep, typeRep )
+import GHC.Exts       ( IsList( toList ) )
+import GHC.Generics   ( Generic )
 
 -- containers --------------------------
 
 import qualified  Data.Sequence  as  Seq
 
-import Data.Sequence  ( Seq( Empty, (:|>) ), (<|) )
-
--- data-default ------------------------
-
-import Data.Default  ( def )
+import Data.Sequence  ( (<|) )
 
 -- data-textual ------------------------
 
-import Data.Textual  ( Printable( print ), Textual( textual )
-                     , fromString, toString, toText )
+import Data.Textual  ( Parsed( Parsed ), Textual( textual )
+                     , fromString, parseString )
+
+-- deepseq -----------------------------
+
+import Control.DeepSeq  ( NFData )
 
 -- lens --------------------------------
 
-import Control.Lens.Cons   ( unsnoc )
-import Control.Lens.Iso    ( iso )
-import Control.Lens.Lens   ( Lens', lens )
-import Control.Lens.Prism  ( Prism', prism' )
+import Control.Lens.Cons    ( unsnoc )
+import Control.Lens.Getter  ( view )
+import Control.Lens.Iso     ( iso )
 
 -- monaderror-io -----------------------
 
@@ -75,31 +58,26 @@ import Data.MonoTraversable  ( Element, MonoFoldable( ofoldl', ofoldl1Ex'
 
 -- more-unicode ------------------------
 
-import Data.MoreUnicode.Applicative  ( (∤), (⋪) )
-import Data.MoreUnicode.Function     ( (⅋) )
-import Data.MoreUnicode.Functor      ( (⊳), (⩺) )
-import Data.MoreUnicode.Lens         ( (⊣), (⊩), (⊢) )
-import Data.MoreUnicode.Monoid       ( ф )
-import Data.MoreUnicode.Natural      ( ℕ )
-
--- mtl ---------------------------------
-
-import Control.Monad.Except  ( MonadError )
+import Data.MoreUnicode.Function  ( (⅋) )
+import Data.MoreUnicode.Lens      ( (⊩) )
 
 -- non-empty-containers ----------------
 
 import qualified  NonEmptyContainers.SeqNE  as  SeqNE
 
+import NonEmptyContainers.IsNonEmpty        ( fromNonEmpty )
 import NonEmptyContainers.SeqConversions    ( FromSeq( fromSeq )
                                             , IsSeq( seq )
                                             , ToSeq( toSeq ) )
-import NonEmptyContainers.SeqNE             ( pattern(:⪭), (⪫), (⋖) )
-import NonEmptyContainers.SeqNEConversions  ( FromSeqNonEmpty( fromSeqNE ) )
+import NonEmptyContainers.SeqNE             ( SeqNE( (:⫸) ) , pattern(:⪭)
+                                            , pattern(:⪬), (⋖) )
+import NonEmptyContainers.SeqNEConversions  ( FromSeqNonEmpty( fromSeqNE )
+                                            , ToSeqNonEmpty( toSeqNE ) )
 
 -- parsers -----------------------------
 
 import Text.Parser.Char         ( char, string )
-import Text.Parser.Combinators  ( endBy )
+import Text.Parser.Combinators  ( endBy, endByNonEmpty )
 
 -- quasiquoting ------------------------
 
@@ -109,35 +87,29 @@ import QuasiQuoting  ( QuasiQuoter, mkQQ, exp )
 
 import Test.QuickCheck.Arbitrary  ( Arbitrary( arbitrary, shrink ) )
 
--- tasty -------------------------------
-
-import Test.Tasty  ( TestTree, testGroup )
-
--- tasty-hunit -------------------------
-
-import Test.Tasty.HUnit  ( (@=?), testCase )
-
 -- tasty-plus --------------------------
 
-import TastyPlus  ( (≟), assertListEq, runTestsP, runTestsReplay, runTestTree )
+import TastyPlus  ( (≟), assertListEq, propInvertibleString, propInvertibleText
+                  , propInvertibleUtf8 )
+
+-- tasty-quickcheck --------------------
+
+import Test.Tasty.QuickCheck  ( testProperty )
 
 -- template-haskell --------------------
 
 import Language.Haskell.TH         ( ExpQ )
-import Language.Haskell.TH.Syntax  ( Exp( AppE, ConE, VarE )
+import Language.Haskell.TH.Syntax  ( Exp( AppE, ConE )
                                    , Lift( lift, liftTyped ), TExp( TExp ) )
 
 -- text --------------------------------
 
-import Data.Text  ( Text, empty, last, splitOn )
+import qualified  Data.Text  as  Text
+import Data.Text  ( Text, empty, splitOn )
 
 -- text-printer ------------------------
 
 import qualified  Text.Printer  as  P
-
--- tfmt --------------------------------
-
-import Text.Fmt  ( fmt )
 
 ------------------------------------------------------------
 --                     local imports                      --
@@ -157,11 +129,13 @@ import FPath.Error.FPathComponentError
 import FPath.Error.FPathError  ( AsFPathError, FPathError( FPathComponentE )
                                , __FPathComponentE__, __FPathEmptyE__
                                , __FPathAbsE__, __FPathNotADirE__
+                               , __FPathRootDirE__
                                , _FPathComponentE
                                , fPathAbsE, fPathEmptyE
-                               , fPathNotADirE
+                               , fPathNotADirE, mapTypeRepE
                                )
-import FPath.Parent            ( HasParentMay( parentMay, parents ) )
+import FPath.Parent            ( HasParent( parent )
+                               , HasParentMay( parentMay, parents ) )
 import FPath.Parseable         ( Parseable( parse ) )
 import FPath.PathComponent     ( PathComponent, parsePathC, pc, toUpper )
 import FPath.RelType           ( RelTypeC( RelType ) )
@@ -169,27 +143,220 @@ import FPath.Util              ( __ERROR'__ )
 
 -------------------------------------------------------------------------------
 
-{- | a relative directory -}
-newtype RelDir = RelDir (Seq PathComponent)
-  deriving newtype (Eq, Monoid, Semigroup)
+{- | A non-root absolute directory, e.g., /etc. -}
+-- a non-root dir is a path component appended to a (possibly-root) absolute
+-- directory
+newtype NonRootRelDir = NonRootRelDir (SeqNE PathComponent)
+  deriving (Eq,Generic,Lift)
+  deriving anyclass NFData
 
-instance Show RelDir where
+nrreldirT ∷ TypeRep
+nrreldirT = typeRep (Proxy ∷ Proxy NonRootRelDir)
+
+type instance Element NonRootRelDir = PathComponent
+
+instance Show NonRootRelDir where
   show r = [fmt|[reldir|%T%s]|] (toText r) "|"
 
-instance Lift RelDir where
-  liftTyped (RelDir ps) = do
-    xs ← lift $ toList ps
-    return ∘ TExp $ AppE (ConE 'RelDir) (AppE (VarE 'fromList) xs)
+----------------------------------------
+
+instance Printable NonRootRelDir where
+  print (NonRootRelDir ps) = pDir id ps
+
+--------------------
+
+instance Textual NonRootRelDir where
+  textual = fromSeqNE ∘ fromNonEmpty ⊳ (endByNonEmpty textual (char '/'))
+
+--------------------
+
+instance AsFilePath NonRootRelDir where
+  filepath = prism' toString fromString
+
+--------------------
+
+instance AsFilePath' NonRootRelDir where
+  filepath' = prism' (exterminate ∘ toString)
+                     (fromString ∘ terminate)
+
+--------------------
+
+instance Semigroup NonRootRelDir where
+  (NonRootRelDir ps) <> (NonRootRelDir ps') = NonRootRelDir $ ps ◇ ps'
+
+--------------------
+
+instance RelTypeC NonRootRelDir where
+  type RelType NonRootRelDir = NonRootRelDir
+
+--------------------
+
+instance DirTypeC NonRootRelDir where
+  type DirType NonRootRelDir = RelDir
+
+----------------------------------------
+
+instance FromSeqNonEmpty NonRootRelDir where
+  fromSeqNE = NonRootRelDir ∘ fromSeqNE
+
+----------------------------------------
+
+instance ToSeqNonEmpty NonRootRelDir where
+  toSeqNE (NonRootRelDir d) = d
+
+----------------------------------------
+
+instance ToSeq NonRootRelDir where
+  toSeq (NonRootRelDir ps) = toSeq ps
+
+----------------------------------------
+
+instance Basename NonRootRelDir where
+  basename (NonRootRelDir (_ :⫸ p)) = fromSeqNE (pure p)
+  updateBasename f (NonRootRelDir (ps :⫸ p)) = NonRootRelDir (ps :⪭ f p)
+
+----------------------------------------
+
+nrrdGetParent ∷ NonRootRelDir → RelDir
+nrrdGetParent (NonRootRelDir (ps :⫸ _)) = fromSeq ps
+
+nrrdSetParent ∷ NonRootRelDir → RelDir → NonRootRelDir
+nrrdSetParent (NonRootRelDir (_ :⫸ d)) p = NonRootRelDir (toSeq p :⫸ d)
+
+instance HasParent NonRootRelDir where
+  parent = lens nrrdGetParent nrrdSetParent
+----------------------------------------
+
+instance HasDirname NonRootRelDir where
+  dirname ∷ Lens' NonRootRelDir RelDir
+  dirname = lens nrrdGetParent nrrdSetParent
+
+  ancestors' ∷ NonRootRelDir → [RelDir]
+  ancestors' d = (d ⊣ dirname) : ancestors' (d ⊣ dirname)
+
+----------------------------------------
+
+nonRootGetParentMay ∷ NonRootRelDir → Maybe RelDir
+nonRootGetParentMay = 𝕵 ∘ view parent
+
+nonRootSetParentMay ∷ NonRootRelDir → Maybe RelDir → NonRootRelDir
+nonRootSetParentMay (NonRootRelDir ps) d =
+  fromSeqNE $ (maybe ф toSeq d) :⫸ SeqNE.last ps
+
+instance HasParentMay NonRootRelDir where
+  parentMay = lens nonRootGetParentMay nonRootSetParentMay
+
+----------------------------------------
+
+instance MonoFunctor NonRootRelDir where
+  omap ∷ (PathComponent → PathComponent) → NonRootRelDir → NonRootRelDir
+  omap f (NonRootRelDir ps) = NonRootRelDir (omap f ps)
+
+----------------------------------------
+
+instance MonoFoldable NonRootRelDir where
+  otoList (NonRootRelDir ps) = toList $ toSeq ps
+  ofoldl' ∷ (α → PathComponent → α) → α → NonRootRelDir → α
+  ofoldl' f x r = foldl' f x (otoList r)
+  ofoldr ∷ (PathComponent → α → α) → α → NonRootRelDir → α
+  ofoldr f x r = foldr f x (otoList r)
+  ofoldMap ∷ Monoid ν ⇒ (PathComponent → ν) → NonRootRelDir → ν
+  ofoldMap f r = foldMap f (otoList r)
+  ofoldr1Ex ∷ (PathComponent → PathComponent → PathComponent) → NonRootRelDir
+            → PathComponent
+  ofoldr1Ex f r = foldr1 f (otoList r)
+  ofoldl1Ex' ∷ (PathComponent → PathComponent → PathComponent) → NonRootRelDir
+             → PathComponent
+  ofoldl1Ex' f r = foldl1 f (otoList r)
+
+----------------------------------------
+
+instance Parseable NonRootRelDir where
+  parse ∷ (AsFPathError ε, MonadError ε η, Printable τ) ⇒ τ → η NonRootRelDir
+  parse t = do mapTypeRepE (const nrreldirT) $ parse t ≫ \ case
+                 RelRootDir → __FPathRootDirE__ nrreldirT
+                 RelNonRootDir d' → return d'
+
+------------------------------------------------------------
+--                     Quasi-Quoting                      --
+------------------------------------------------------------
+
+reldirNQQ ∷ String → Maybe ExpQ
+reldirNQQ = (\ d → ⟦d⟧) ⩺ (ѭ ∘ parse @NonRootRelDir @FPathError)
+
+reldirN ∷ QuasiQuoter
+reldirN = mkQQ "NonRootRelDir" $ def & exp ⊩ reldirNQQ
+
+------------------------------------------------------------
+
+{- | a relative directory -}
+data RelDir = RelRootDir | RelNonRootDir NonRootRelDir
+  deriving Eq
+
+--------------------
 
 type instance Element RelDir = PathComponent
 
 --------------------
 
+instance Show RelDir where
+  show r = [fmt|[reldir|%T%s]|] (toText r) "|"
+
+showTests ∷ TestTree
+showTests =
+  testGroup "show"
+            [ testCase "root"  $ "[reldir|./|]"           ≟ show r0
+            , testCase "etc"   $ "[reldir|r/|]"       ≟ show r1
+            , testCase "pam.d" $ "[reldir|r/p/|]" ≟ show r2
+            , testCase "pam.d" $ "[reldir|p/q/r/|]" ≟ show r3
+            , testCase "pam.d" $ "[reldir|p/|]" ≟ show r3p
+            , testCase "pam.d" $ "[reldir|p/q/|]" ≟ show r3pq
+            ]
+
+--------------------
+
+instance Lift RelDir where
+  liftTyped RelRootDir = return ∘ TExp $ ConE 'RelRootDir
+  liftTyped (RelNonRootDir d) = do
+    x ← lift d
+    return ∘ TExp $ AppE (ConE 'RelNonRootDir) x
+
+--------------------
+
+instance Semigroup RelDir where
+  RelRootDir <> RelRootDir = RelRootDir
+  RelRootDir <> r@(RelNonRootDir _) = r
+  r@(RelNonRootDir _) <> RelRootDir = r
+  (RelNonRootDir r) <> (RelNonRootDir r') = RelNonRootDir $ r ◇ r'
+
+--------------------
+
+instance Monoid RelDir where
+  mempty = RelRootDir
+
+--------------------
+
+{-| Things that may convert to an `RelDir` (but  a `RelDir` will always convert
+    to); e.g., @Dir@, @Rel@, @FPath@. -}
 class AsRelDir α where
   _RelDir ∷ Prism' α RelDir
 
 instance AsRelDir RelDir where
   _RelDir = id
+
+--------------------
+
+{-| Things that /may/ be converted from an `RelDir` (but will always convert
+    /to/ an `RelDir`). -}
+class RelDirAs α where
+  _RelDir_ ∷ Prism' RelDir α
+
+instance RelDirAs RelDir where
+  _RelDir_ = id
+
+instance RelDirAs NonRootRelDir where
+  _RelDir_ =
+    prism' RelNonRootDir (\ case RelRootDir → 𝕹; RelNonRootDir d → 𝕵 d)
 
 --------------------
 
@@ -205,19 +372,21 @@ instance RelTypeC RelDir where
 
 instance MonoFunctor RelDir where
   omap ∷ (PathComponent → PathComponent) → RelDir → RelDir
-  omap f (RelDir ps) = RelDir (omap f ps)
+  omap _ RelRootDir = RelRootDir
+  omap f (RelNonRootDir ps) = RelNonRootDir (omap f ps)
 
 ----------------------------------------
 
 instance MonoFoldable RelDir where
   otoList ∷ RelDir → [PathComponent]
-  otoList (RelDir ps) = toList ps
+  otoList RelRootDir = ф
+  otoList (RelNonRootDir d) = otoList d
   ofoldl' ∷ (α → PathComponent → α) → α → RelDir → α
   ofoldl' f x r = foldl' f x (toList r)
 
   ofoldr ∷ (PathComponent → α → α) → α → RelDir → α
   ofoldr f x r = foldr f x (toList r)
-  ofoldMap ∷ Monoid ν => (PathComponent → ν) → RelDir → ν
+  ofoldMap ∷ Monoid ν ⇒ (PathComponent → ν) → RelDir → ν
   ofoldMap f r = foldMap f (toList r)
   ofoldr1Ex ∷ (PathComponent → PathComponent → PathComponent) → RelDir
             → PathComponent
@@ -229,17 +398,20 @@ instance MonoFoldable RelDir where
 ----------------------------------------
 
 instance FromSeqNonEmpty RelDir where
-  fromSeqNE = RelDir ∘ SeqNE.toSeq
+  fromSeqNE = RelNonRootDir ∘ fromSeqNE
 
 ----------------------------------------
 
 instance FromSeq RelDir where
-  fromSeq = RelDir
+  fromSeq Seq.Empty = RelRootDir
+  fromSeq (x :⪬ xs) = RelNonRootDir $ fromSeqNE (x :⪬ xs)
+  fromSeq _         = error "patterns should not be exhausted!"
 
 ----------------------------------------
 
 instance ToSeq RelDir where
-  toSeq (RelDir ps) = ps
+  toSeq RelRootDir        = ф
+  toSeq (RelNonRootDir d) = toSeq d
 
 ----------------------------------------
 
@@ -250,7 +422,8 @@ instance IsSeq RelDir where
 
 instance IsList RelDir where
   type instance Item RelDir = PathComponent
-  fromList = RelDir ∘ Seq.fromList
+  fromList []     = RelRootDir
+  fromList (x:xs) = RelNonRootDir $ fromSeqNE (x ⋖ xs)
   toList   = toList ∘ toSeq
 
 ----------------------------------------
@@ -262,8 +435,8 @@ pDir ∷ (P.Printer ρ, ToSeq α, Printable (Element α)) ⇒ (String → String
 pDir f =  P.string ∘ f ∘ concat ∘ fmap ((⊕ "/") ∘ toString) ∘ toSeq
 
 instance Printable RelDir where
-  print (RelDir ps) | ps ≡ ф = "./"
-                    | otherwise = pDir id ps
+  print RelRootDir        = "./"
+  print (RelNonRootDir d) = print d
 
 ----------------------------------------
 
@@ -282,6 +455,38 @@ instance Textual RelDir where
   textual = return (fromList []) ⋪ (string "./")
           ∤ fromList ⊳ (endBy textual (char '/'))
 
+----------
+
+textualTests ∷ TestTree
+textualTests =
+  let success e s = testCase s $ Parsed e                @=? parseString s
+      fail s      = testCase s $ (𝕹 ∷ 𝕄 RelDir)        @=? fromString s
+      failN s     = testCase s $ (𝕹 ∷ 𝕄 NonRootRelDir) @=? fromString s
+   in testGroup "Textual"
+                [ success r0 "./"
+                , success r1 "r/"
+                , success r2 "r/p/"
+                , success r3 "p/q/r/"
+                , success r3p "p/"
+                , success r3pq "p/q/"
+                , success r3pN "p/"
+                , success r3pqN "p/q/"
+                , failN "./"
+                , fail "/etc"
+                , fail "/etc/pam.d"
+                , fail "/etc/"
+                , fail "etc/pam.d"
+                , fail "/etc//pam.d/"
+                , fail "e/c"
+                , fail "\0etc"
+                , fail "etc\0"
+                , fail "e\0c"
+                , testProperty "parseString - toString"
+                               (propInvertibleString @RelDir)
+                , testProperty "parseText - toText" (propInvertibleText @RelDir)
+                , testProperty "parseUtf8 - toUtf8" (propInvertibleUtf8 @RelDir)
+                ]
+
 ----------------------------------------
 
 instance Arbitrary RelDir where
@@ -293,18 +498,14 @@ instance Arbitrary RelDir where
 instance HasParentMay RelDir where
   parentMay = lens getParentMay setParentMay
               where getParentMay ∷ RelDir → Maybe RelDir
-                    getParentMay (RelDir ps) = case unsnoc ps of
-                                                 Just (p,_) → Just $ RelDir p
-                                                 Nothing → Nothing
+                    getParentMay RelRootDir = 𝕹
+                    getParentMay (RelNonRootDir d) = d ⊣ parentMay
 
-                    setParentMay (RelDir ps) par =
-                      case unsnoc ps of
-                        Just (_, d) → case par of
-                                        Just (RelDir p) → RelDir $ p ⪫ d
-                                        Nothing         → RelDir $ pure d
-                        Nothing     → case par of
-                                        Just r → r
-                                        Nothing → RelDir Seq.Empty
+                    setParentMay ∷ RelDir → Maybe RelDir → RelDir
+                    setParentMay RelRootDir (𝕵 r) = r
+                    setParentMay RelRootDir 𝕹    = RelRootDir
+                    setParentMay (RelNonRootDir d) r =
+                      RelNonRootDir $ d & parentMay ⊢ r
 
 ----------
 
@@ -323,12 +524,12 @@ parentsTests =
 
 instance Basename RelDir where
   basename ∷ RelDir → RelDir
-  basename (RelDir (_ Seq.:|> p)) = fromList [p]
-  basename r@(RelDir Seq.Empty) = r
+  basename (RelNonRootDir d) = RelNonRootDir $ basename d
+  basename RelRootDir = RelRootDir
 
   updateBasename ∷ (PathComponent → PathComponent) → RelDir → RelDir
-  updateBasename f (RelDir (ps Seq.:|> p)) = RelDir (ps :⪭ f p)
-  updateBasename _ r@(RelDir Seq.Empty) = r
+  updateBasename f (RelNonRootDir d) = RelNonRootDir $ updateBasename f d
+  updateBasename _ RelRootDir = RelRootDir
 
 basenameTests ∷ TestTree
 basenameTests =
@@ -356,16 +557,16 @@ basenameTests =
 
 instance HasDirname RelDir where
   dirname ∷ Lens' RelDir RelDir
-  dirname = lens (\ case RelDir Empty     → RelDir Empty
-                         RelDir (s :|> _) → RelDir s)
-                 (\ r (RelDir d) → case r of
-                                     RelDir Empty     → RelDir d
-                                     RelDir (_ :|> a) → RelDir (d :|> a)
+  dirname = lens (\ case RelRootDir      → RelRootDir
+                         RelNonRootDir d → d ⊣ dirname)
+                 (\ r d → case r of
+                            RelRootDir       → d
+                            RelNonRootDir r' → RelNonRootDir $ r' & dirname ⊢ d
                  )
 
   ancestors' ∷ RelDir → [RelDir]
-  ancestors' (RelDir Empty) = []
-  ancestors' fp             = (fp ⊣ dirname) : ancestors' (fp ⊣ dirname)
+  ancestors' RelRootDir = []
+  ancestors' fp         = (fp ⊣ dirname) : ancestors' (fp ⊣ dirname)
 
 dirnameTests ∷ TestTree
 dirnameTests =
@@ -423,11 +624,11 @@ instance Parseable RelDir where
   parse ∷ (AsFPathError ε, MonadError ε η, Printable τ) ⇒ τ → η RelDir
   parse (toText → t) =
     case unsnoc $ splitOn "/" t of
-      Nothing           → error "cannot happen: splitOn always returns something"
-      Just (("":_), _)  → __FPathAbsE__ reldirT t
-      Just ([],"")      → __FPathEmptyE__ reldirT
-      Just (["."],"")   → return $ RelDir ф
-      Just ((x:xs), "") → do
+      𝕹           → error "cannot happen: splitOn always returns something"
+      𝕵 (("":_), _)  → __FPathAbsE__ reldirT t
+      𝕵 ([],"")      → __FPathEmptyE__ reldirT
+      𝕵 (["."],"")   → return RelRootDir
+      𝕵 ((x:xs), "") → do
         let mkCompE ∷ (AsFPathError ε', MonadError ε' η') ⇒
                       FPathComponentError → η' α
             mkCompE ce = __FPathComponentE__ ce reldirT t
@@ -437,7 +638,7 @@ instance Parseable RelDir where
 
         p  ← eCompE $ parsePathC x
         ps ← eCompE $ mapM parsePathC xs
-        return $ RelDir (p <| Seq.fromList ps)
+        return $ fromSeq (p <| Seq.fromList ps)
       _                 → __FPathNotADirE__ reldirT t
 
 --------------------
@@ -448,24 +649,24 @@ parseRelDirTests =
       illegalCE s t = let fpcice = fPathComponentIllegalCharE '\0' t
                        in FPathComponentE fpcice reldirT s
       badChar s p = testCase ("bad component " ⊕ toString s) $
-                        Left (illegalCE s p) @=? parseRelDir_ s
+                        𝕷 (illegalCE s p) @=? parseRelDir_ s
       emptyCompCE t = FPathComponentE fPathComponentEmptyE reldirT t
       parseRelDir_ ∷ MonadError FPathError η ⇒ Text → η RelDir
       parseRelDir_ = parse
    in testGroup "parseRelDir"
-                [ testCase "r0 (./)" $ Right r0 @=? parseRelDir_ "./"
-                , testCase "r1" $ Right r1 @=? parseRelDir_ "r/"
-                , testCase "r2" $ Right r2 @=? parseRelDir_ "r/p/"
-                , testCase "r3" $ Right r3 @=? parseRelDir_ "p/q/r/"
+                [ testCase "r0 (./)" $ 𝕽 r0 @=? parseRelDir_ "./"
+                , testCase "r1" $ 𝕽 r1 @=? parseRelDir_ "r/"
+                , testCase "r2" $ 𝕽 r2 @=? parseRelDir_ "r/p/"
+                , testCase "r3" $ 𝕽 r3 @=? parseRelDir_ "p/q/r/"
                 , testCase "no trailing /" $
-                      Left (fPathNotADirE reldirT pamF) @=? parseRelDir_ pamF
+                      𝕷 (fPathNotADirE reldirT pamF) @=? parseRelDir_ pamF
                 , testCase "leading /" $
-                      Left (fPathAbsE reldirT "/r/") @=? parseRelDir_ "/r/"
+                      𝕷 (fPathAbsE reldirT "/r/") @=? parseRelDir_ "/r/"
                 , badChar "x/\0/y/" "\0"
                 , badChar "r/p\0/" "p\0"
                 , badChar "\0r/p/" "\0r"
                 , testCase "empty component" $
-                      Left (emptyCompCE "r//p/") @=? parseRelDir_ "r//p/"
+                      𝕷 (emptyCompCE "r//p/") @=? parseRelDir_ "r//p/"
                 ]
 
 ----------------------------------------
@@ -475,11 +676,11 @@ parseRelDirTests =
      `parseRelDir`" -}
 parseRelDirP ∷ (AsFPathError ε, MonadError ε η, Printable τ) ⇒ τ → η RelDir
 parseRelDirP (toText → t) =
-  let safeLast "" = Nothing
-      safeLast s  = Just $ last s
+  let safeLast "" = 𝕹
+      safeLast s  = 𝕵 $ Text.last s
    in case safeLast t of
-        Nothing  → parse empty
-        Just '/' → parse t
+        𝕹  → parse empty
+        𝕵 '/' → parse t
         _        → parse (t ⊕ "/")
 
 parseRelDirP' ∷ (Printable τ, MonadError FPathError η) ⇒ τ → η RelDir
@@ -502,21 +703,21 @@ parseRelDirPTests =
       _parseRelDirP ∷ MonadError FPathError η ⇒ Text → η RelDir
       _parseRelDirP = parseRelDirP'
    in testGroup "parseRelDirP"
-                [ testCase "r0" $ Right r0 @=? _parseRelDirP "."
-                , testCase "r1" $ Right r1 @=? _parseRelDirP "r/"
-                , testCase "r1" $ Right r1 @=? _parseRelDirP "r"
-                , testCase "r2" $ Right r2 @=? _parseRelDirP "r/p/"
-                , testCase "r2" $ Right r2 @=? _parseRelDirP "r/p"
-                , testCase "r3" $ Right r3 @=? _parseRelDirP "p/q/r/"
-                , testCase "r3" $ Right r3 @=? _parseRelDirP "p/q/r"
+                [ testCase "r0" $ 𝕽 r0 @=? _parseRelDirP "."
+                , testCase "r1" $ 𝕽 r1 @=? _parseRelDirP "r/"
+                , testCase "r1" $ 𝕽 r1 @=? _parseRelDirP "r"
+                , testCase "r2" $ 𝕽 r2 @=? _parseRelDirP "r/p/"
+                , testCase "r2" $ 𝕽 r2 @=? _parseRelDirP "r/p"
+                , testCase "r3" $ 𝕽 r3 @=? _parseRelDirP "p/q/r/"
+                , testCase "r3" $ 𝕽 r3 @=? _parseRelDirP "p/q/r"
                 , testCase "empty" $
-                      Left (fPathEmptyE reldirT)  @=? _parseRelDirP ""
+                      𝕷 (fPathEmptyE reldirT)  @=? _parseRelDirP ""
                 , testCase "no leading /" $
-                      Left (fPathAbsE reldirT "/etc/") @=? _parseRelDirP "/etc/"
+                      𝕷 (fPathAbsE reldirT "/etc/") @=? _parseRelDirP "/etc/"
                 , testCase "bad component" $
-                      Left illegalCE @=? _parseRelDirP pamNUL
+                      𝕷 illegalCE @=? _parseRelDirP pamNUL
                 , testCase "empty component" $
-                      Left emptyCompCE @=? _parseRelDirP "etc//pam.d/"
+                      𝕷 emptyCompCE @=? _parseRelDirP "etc//pam.d/"
                 ]
 
 ----------------------------------------
@@ -552,6 +753,12 @@ r3p = fromSeqNE $ pure [pc|p|]
 r3pq ∷ RelDir
 r3pq = fromSeqNE $ [pc|p|] ⋖ [[pc|q|]]
 
+r3pN ∷ NonRootRelDir
+r3pN = fromSeqNE $ pure [pc|p|]
+
+r3pqN ∷ NonRootRelDir
+r3pqN = fromSeqNE $ [pc|p|] ⋖ [[pc|q|]]
+
 ----------------------------------------
 
 constructionTests ∷ TestTree
@@ -562,7 +769,8 @@ constructionTests = testGroup "construction" [ parseRelDirTests
 tests ∷ TestTree
 tests = testGroup "FPath.RelDir"
                   [ constructionTests, basenameTests, dirnameTests
-                  , parentsTests, ancestors'Tests ]
+                  , parentsTests, ancestors'Tests, showTests, textualTests
+                  ]
 
 ----------------------------------------